Split project loader tests from workspace model
