+++ Tests/QuillCodeAppTests/WorkspaceModelTests.swift
@@ -2574,106 +2574,6 @@ final class WorkspaceModelTests: XCTestCase {
         XCTAssertTrue(model.selectedThread?.instructions.first?.content.contains("targeted unit tests") == true)
     }
 
-    func testProjectInstructionLoaderBoundsFilesAndRejectsSymlinkEscape() throws {
-        let root = try makeTempDirectory()
-        let outside = try makeTempDirectory().appendingPathComponent("outside.md")
-        try "outside rules\n".write(to: outside, atomically: true, encoding: .utf8)
-        try FileManager.default.createSymbolicLink(
-            at: root.appendingPathComponent("AGENTS.md"),
-            withDestinationURL: outside
-        )
-        let quillcodeDirectory = root.appendingPathComponent(".quillcode")
-        try FileManager.default.createDirectory(at: quillcodeDirectory, withIntermediateDirectories: true)
-        try String(repeating: "x", count: 64).write(
-            to: quillcodeDirectory.appendingPathComponent("rules.md"),
-            atomically: true,
-            encoding: .utf8
-        )
-
-        let instructions = ProjectInstructionLoader.load(
-            from: root,
-            maxFileBytes: 12,
-            maxTotalBytes: 20
-        )
-
-        XCTAssertEqual(instructions.map(\.path), [".quillcode/rules.md"])
-        XCTAssertTrue(instructions[0].wasTruncated)
-        XCTAssertTrue(instructions[0].content.contains("truncated"))
-        XCTAssertFalse(instructions[0].content.contains("outside rules"))
-    }
-
-    func testProjectInstructionLoaderLoadsNestedInstructionsInPrecedenceOrder() throws {
-        let root = try makeTempDirectory()
-        try "Root rules\n".write(
-            to: root.appendingPathComponent("AGENTS.md"),
-            atomically: true,
-            encoding: .utf8
-        )
-
-        let feature = root.appendingPathComponent("Sources/Feature")
-        try FileManager.default.createDirectory(at: feature, withIntermediateDirectories: true)
-        try "Sources rules\n".write(
-            to: root.appendingPathComponent("Sources/AGENTS.md"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try "Feature rules\n".write(
-            to: feature.appendingPathComponent("AGENTS.md"),
-            atomically: true,
-            encoding: .utf8
-        )
-
-        let featureQuillCode = feature.appendingPathComponent(".quillcode")
-        try FileManager.default.createDirectory(at: featureQuillCode, withIntermediateDirectories: true)
-        try "Feature QuillCode rules\n".write(
-            to: featureQuillCode.appendingPathComponent("rules.md"),
-            atomically: true,
-            encoding: .utf8
-        )
-
-        let generated = root.appendingPathComponent(".build/generated")
-        try FileManager.default.createDirectory(at: generated, withIntermediateDirectories: true)
-        try "Generated rules should not load\n".write(
-            to: generated.appendingPathComponent("AGENTS.md"),
-            atomically: true,
-            encoding: .utf8
-        )
-
-        let instructions = ProjectInstructionLoader.load(from: root)
-
-        XCTAssertEqual(instructions.map(\.path), [
-            "AGENTS.md",
-            "Sources/AGENTS.md",
-            "Sources/Feature/AGENTS.md",
-            "Sources/Feature/.quillcode/rules.md"
-        ])
-        XCTAssertTrue(instructions.last?.content.contains("Feature QuillCode rules") == true)
-        XCTAssertFalse(instructions.contains { $0.content.contains("Generated rules") })
-    }
-
-    func testProjectInstructionLoaderCapsNestedInstructionCount() throws {
-        let root = try makeTempDirectory()
-        for index in 0..<5 {
-            let directory = root.appendingPathComponent("Area\(index)")
-            try FileManager.default.createDirectory(at: directory, withIntermediateDirectories: true)
-            try "Rules \(index)\n".write(
-                to: directory.appendingPathComponent("AGENTS.md"),
-                atomically: true,
-                encoding: .utf8
-            )
-        }
-
-        let instructions = ProjectInstructionLoader.load(
-            from: root,
-            maxInstructionFiles: 2
-        )
-
-        XCTAssertEqual(instructions.map(\.path), [
-            "Area0/AGENTS.md",
-            "Area1/AGENTS.md"
-        ])
-    }
-
     func testLocalEnvironmentActionsLoadAndRunFromCommandPaletteIDs() throws {
         let root = try makeTempDirectory()
         let actionsDirectory = root.appendingPathComponent(".quillcode/actions")
@@ -2699,57 +2599,6 @@ final class WorkspaceModelTests: XCTestCase {
         let outputJSON = try XCTUnwrap(card.outputJSON)
         let result = try JSONHelpers.decode(ToolResult.self, from: outputJSON)
         XCTAssertEqual(result.stdout, "local-env-ok")
-    }
-
-    func testLocalEnvironmentActionLoaderUsesMetadataSidecars() throws {
-        let root = try makeTempDirectory()
-        let actionsDirectory = root.appendingPathComponent(".quillcode/actions")
-        try FileManager.default.createDirectory(at: actionsDirectory, withIntermediateDirectories: true)
-        try "printf second".write(
-            to: actionsDirectory.appendingPathComponent("z-second.sh"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try """
-        {
-          "title": "Second Check",
-          "description": "Runs after dependencies are ready.",
-          "order": 20
-        }
-        """.write(
-            to: actionsDirectory.appendingPathComponent("z-second.json"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try "printf first".write(
-            to: actionsDirectory.appendingPathComponent("a-first.sh"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try """
-        {
-          "title": "Prepare Workspace",
-          "description": "Install dependencies and warm caches.",
-          "order": 10
-        }
-        """.write(
-            to: actionsDirectory.appendingPathComponent("a-first.json"),
-            atomically: true,
-            encoding: .utf8
-        )
-
-        let actions = LocalEnvironmentActionLoader.load(from: root)
-
-        XCTAssertEqual(actions.map(\.title), ["Prepare Workspace", "Second Check"])
-        XCTAssertEqual(actions.map(\.detail), [
-            "Install dependencies and warm caches.",
-            "Runs after dependencies are ready."
-        ])
-        XCTAssertEqual(actions.map(\.relativePath), [
-            ".quillcode/actions/a-first.sh",
-            ".quillcode/actions/z-second.sh"
-        ])
-        XCTAssertEqual(actions[0].command, #"sh '.quillcode/actions/a-first.sh'"#)
     }
 
     func testLocalEnvironmentActionMetadataInjectsBoundedEnvironment() throws {
@@ -2884,62 +2733,6 @@ final class WorkspaceModelTests: XCTestCase {
         XCTAssertEqual(result.error, "Command timed out after 1s.")
     }
 
-    func testLocalEnvironmentActionLoaderRejectsUnsafeWorkingDirectory() throws {
-        let root = try makeTempDirectory()
-        let outsideDirectory = try makeTempDirectory()
-        let actionsDirectory = root.appendingPathComponent(".quillcode/actions")
-        try FileManager.default.createDirectory(at: actionsDirectory, withIntermediateDirectories: true)
-        try FileManager.default.createSymbolicLink(
-            at: root.appendingPathComponent("escape"),
-            withDestinationURL: outsideDirectory
-        )
-        try "printf safe".write(
-            to: actionsDirectory.appendingPathComponent("safe.sh"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try """
-        {
-          "title": "Safe",
-          "workingDirectory": "escape"
-        }
-        """.write(
-            to: actionsDirectory.appendingPathComponent("safe.json"),
-            atomically: true,
-            encoding: .utf8
-        )
-
-        let action = try XCTUnwrap(LocalEnvironmentActionLoader.load(from: root).first)
-
-        XCTAssertNil(action.workingDirectory)
-        XCTAssertEqual(action.command, #"sh '.quillcode/actions/safe.sh'"#)
-    }
-
-    func testLocalEnvironmentActionLoaderRejectsUnsafeTimeoutSeconds() throws {
-        let root = try makeTempDirectory()
-        let actionsDirectory = root.appendingPathComponent(".quillcode/actions")
-        try FileManager.default.createDirectory(at: actionsDirectory, withIntermediateDirectories: true)
-        try "printf safe".write(
-            to: actionsDirectory.appendingPathComponent("safe.sh"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try """
-        {
-          "title": "Safe",
-          "timeout_seconds": 1801
-        }
-        """.write(
-            to: actionsDirectory.appendingPathComponent("safe.json"),
-            atomically: true,
-            encoding: .utf8
-        )
-
-        let action = try XCTUnwrap(LocalEnvironmentActionLoader.load(from: root).first)
-
-        XCTAssertNil(action.timeoutSeconds)
-    }
-
     func testSlashEnvironmentActionListShowsMetadataDescription() async throws {
         let root = try makeTempDirectory()
         let actionsDirectory = root.appendingPathComponent(".quillcode/actions")
@@ -2970,163 +2763,6 @@ final class WorkspaceModelTests: XCTestCase {
         let message = try XCTUnwrap(model.selectedThread?.messages.last?.content)
         XCTAssertTrue(message.contains("/env Prepare Workspace"))
         XCTAssertTrue(message.contains("Install dependencies and warm caches."))
-    }
-
-    func testLocalEnvironmentActionLoaderBoundsScriptsAndRejectsSymlinkEscape() throws {
-        let root = try makeTempDirectory()
-        let outside = try makeTempDirectory().appendingPathComponent("outside.sh")
-        try "printf bad".write(to: outside, atomically: true, encoding: .utf8)
-        let outsideMetadata = outside.deletingPathExtension().appendingPathExtension("json")
-        try """
-        { "title": "Escaped Metadata" }
-        """.write(to: outsideMetadata, atomically: true, encoding: .utf8)
-        let actionsDirectory = root.appendingPathComponent(".quillcode/actions")
-        try FileManager.default.createDirectory(at: actionsDirectory, withIntermediateDirectories: true)
-        try FileManager.default.createSymbolicLink(
-            at: actionsDirectory.appendingPathComponent("outside.sh"),
-            withDestinationURL: outside
-        )
-        try "printf one".write(
-            to: actionsDirectory.appendingPathComponent("one.sh"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try FileManager.default.createSymbolicLink(
-            at: actionsDirectory.appendingPathComponent("one.json"),
-            withDestinationURL: outsideMetadata
-        )
-        try "printf two".write(
-            to: actionsDirectory.appendingPathComponent("two.sh"),
-            atomically: true,
-            encoding: .utf8
-        )
-
-        let actions = LocalEnvironmentActionLoader.load(from: root, maxActions: 1)
-
-        XCTAssertEqual(actions.map(\.relativePath), [".quillcode/actions/one.sh"])
-        XCTAssertEqual(actions[0].title, "One")
-        XCTAssertEqual(actions[0].command, #"sh '.quillcode/actions/one.sh'"#)
-    }
-
-    func testProjectExtensionManifestLoaderLoadsKindsAndRejectsUnsafeFiles() throws {
-        let root = try makeTempDirectory()
-        let pluginDirectory = root.appendingPathComponent(".quillcode/plugins")
-        let skillDirectory = root.appendingPathComponent(".quillcode/skills")
-        let mcpDirectory = root.appendingPathComponent(".quillcode/mcp")
-        try FileManager.default.createDirectory(at: pluginDirectory, withIntermediateDirectories: true)
-        try FileManager.default.createDirectory(at: skillDirectory, withIntermediateDirectories: true)
-        try FileManager.default.createDirectory(at: mcpDirectory, withIntermediateDirectories: true)
-
-        try #"{"id":"github","name":"GitHub","description":"PR and issue helpers.","version":"1.2.0","source":"https://github.com/Lore-Hex/quillcode-github","updateCommand":"git -C .quillcode/plugins/github pull --ff-only","updateTimeoutSeconds":300}"#.write(
-            to: pluginDirectory.appendingPathComponent("github.json"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try #"{"id":"review","name":"Code Review","summary":"Review defects first.","enabled":false}"#.write(
-            to: skillDirectory.appendingPathComponent("review.json"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try #"{"id":"filesystem","name":"Filesystem MCP","command":"quill-mcp","args":["--root","."]}"#.write(
-            to: mcpDirectory.appendingPathComponent("filesystem.json"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try #"{"id":"broken""#.write(
-            to: pluginDirectory.appendingPathComponent("broken.json"),
-            atomically: true,
-            encoding: .utf8
-        )
-        let outside = try makeTempDirectory().appendingPathComponent("outside.json")
-        try #"{"id":"outside","name":"Outside"}"#.write(to: outside, atomically: true, encoding: .utf8)
-        try FileManager.default.createSymbolicLink(
-            at: pluginDirectory.appendingPathComponent("outside.json"),
-            withDestinationURL: outside
-        )
-
-        let manifests = ProjectExtensionManifestLoader.load(from: root)
-
-        XCTAssertEqual(manifests.map(\.id), [
-            "plugin:github",
-            "skill:review",
-            "mcp_server:filesystem"
-        ])
-        XCTAssertEqual(manifests.map(\.kind), [.plugin, .skill, .mcpServer])
-        XCTAssertEqual(manifests[0].summary, "PR and issue helpers.")
-        XCTAssertEqual(manifests[0].version, "1.2.0")
-        XCTAssertEqual(manifests[0].sourceURL, "https://github.com/Lore-Hex/quillcode-github")
-        XCTAssertEqual(manifests[0].updateCommand, "git -C .quillcode/plugins/github pull --ff-only")
-        XCTAssertEqual(manifests[0].updateTimeoutSeconds, 300)
-        XCTAssertEqual(manifests[1].isEnabled, false)
-        XCTAssertEqual(manifests[2].transport, .stdio)
-        XCTAssertEqual(manifests[2].launchExecutable, "quill-mcp")
-        XCTAssertEqual(manifests[2].launchCommand, "quill-mcp --root .")
-        XCTAssertEqual(manifests[2].launchArguments, ["--root", "."])
-    }
-
-    func testProjectExtensionManifestLoaderSkipsUnsafeCustomDirectoriesWithoutStoppingScan() throws {
-        let root = try makeTempDirectory()
-        let safeDirectory = root.appendingPathComponent(".quillcode/plugins")
-        try FileManager.default.createDirectory(at: safeDirectory, withIntermediateDirectories: true)
-        try #"{"id":"github","name":"GitHub"}"#.write(
-            to: safeDirectory.appendingPathComponent("github.json"),
-            atomically: true,
-            encoding: .utf8
-        )
-
-        let manifests = ProjectExtensionManifestLoader.load(
-            from: root,
-            directories: [
-                ("../outside", .plugin),
-                ("/absolute", .skill),
-                (".quillcode/./mcp", .mcpServer),
-                (".quillcode/plugins", .plugin)
-            ]
-        )
-
-        XCTAssertEqual(manifests.map(\.id), ["plugin:github"])
-        XCTAssertEqual(manifests.map(\.relativePath), [".quillcode/plugins/github.json"])
-    }
-
-    func testProjectExtensionManifestLoaderSkipsSymlinkedDirectoryOutsideProject() throws {
-        let root = try makeTempDirectory()
-        let quillCodeDirectory = root.appendingPathComponent(".quillcode")
-        let outsideDirectory = try makeTempDirectory().appendingPathComponent("plugins")
-        try FileManager.default.createDirectory(at: quillCodeDirectory, withIntermediateDirectories: true)
-        try FileManager.default.createDirectory(at: outsideDirectory, withIntermediateDirectories: true)
-        try #"{"id":"outside","name":"Outside"}"#.write(
-            to: outsideDirectory.appendingPathComponent("outside.json"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try FileManager.default.createSymbolicLink(
-            at: quillCodeDirectory.appendingPathComponent("plugins"),
-            withDestinationURL: outsideDirectory
-        )
-
-        let manifests = ProjectExtensionManifestLoader.load(from: root)
-
-        XCTAssertTrue(manifests.isEmpty)
-    }
-
-    func testProjectExtensionManifestLoaderFallsBackToFilenameForBlankOrMissingName() throws {
-        let root = try makeTempDirectory()
-        let pluginDirectory = root.appendingPathComponent(".quillcode/plugins")
-        try FileManager.default.createDirectory(at: pluginDirectory, withIntermediateDirectories: true)
-        try #"{"id":"code-review","name":"   "}"#.write(
-            to: pluginDirectory.appendingPathComponent("code-review.json"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try #"{"id":"issue-triage"}"#.write(
-            to: pluginDirectory.appendingPathComponent("issue-triage.json"),
-            atomically: true,
-            encoding: .utf8
-        )
-
-        let manifests = ProjectExtensionManifestLoader.load(from: root)
-
-        XCTAssertEqual(manifests.map(\.name), ["Code Review", "Issue Triage"])
     }
 
     func testProjectExtensionManifestsLoadIntoProjectSurface() throws {
@@ -3620,40 +3256,6 @@ final class WorkspaceModelTests: XCTestCase {
         XCTAssertTrue(model.runWorkspaceCommand("memory-add", workspaceRoot: try makeTempDirectory()))
 
         XCTAssertEqual(model.composer.draft, "/remember ")
-    }
-
-    func testMemoryNoteLoaderBoundsFilesAndRejectsSymlinkEscape() throws {
-        let root = try makeTempDirectory()
-        let outside = try makeTempDirectory().appendingPathComponent("outside.md")
-        try "outside memory\n".write(to: outside, atomically: true, encoding: .utf8)
-        let memoryDirectory = root.appendingPathComponent(".quillcode/memories")
-        try FileManager.default.createDirectory(at: memoryDirectory, withIntermediateDirectories: true)
-        try FileManager.default.createSymbolicLink(
-            at: memoryDirectory.appendingPathComponent("outside.md"),
-            withDestinationURL: outside
-        )
-        try String(repeating: "x", count: 64).write(
-            to: memoryDirectory.appendingPathComponent("one.md"),
-            atomically: true,
-            encoding: .utf8
-        )
-        try "ignored binary".write(
-            to: memoryDirectory.appendingPathComponent("ignored.bin"),
-            atomically: true,
-            encoding: .utf8
-        )
-
-        let notes = MemoryNoteLoader.loadProject(
-            from: root,
-            maxNotes: 1,
-            maxFileBytes: 12,
-            maxTotalBytes: 12
-        )
-
-        XCTAssertEqual(notes.map(\.relativePath), [".quillcode/memories/one.md"])
-        XCTAssertTrue(notes[0].wasTruncated)
-        XCTAssertTrue(notes[0].content.contains("truncated"))
-        XCTAssertFalse(notes[0].content.contains("outside memory"))
     }
 
     func testEmptyDraftDoesNotCreateThread() async throws {
